History (#55) (#56)
* Refactor login check into its own method

Moved the function for checking if a user is logged in into its own method. This is intended to enhance code readability and maintainability. It eliminates duplicated code blocks across multiple files, providing a single point of control for logic adjustments.

* Refactor history browsing function and template parameters

Separated display and data fetching concerns for user history browsing functionality. This heightens code readability and better follows software design principles. Also, optional chaining is added in item templates to ensure the code won't break due to missing or null data.

* Implement separate HistoryModel for user history browsing

The previous implementation had the history functionality within the PeminjamanModel. This commit extracts the history() function and places it into a new HistoryModel, thereby achieving a better separation of concerns. This change enhances the modular approach, improving the code's readability and maintainability.

* Refactor user history feature into separate HistoryModel

This commit moves the user history browsing feature from the PeminjamanModel to a newly created HistoryModel. This refactoring process improves the separation of concerns and promotes modularity, enhancing the code's maintainability and readability.