UI (#57)
* Update breakpoint settings in Tailwind configuration

The Tailwind CSS configuration file was updated to change the breakpoint settings. Instead of setting a min and max range for each breakpoint, it's now using a min-width approach. This aligns with standard CSS media queries, that usually focus on minimum widths.

* Improve flash message handling and refactor login page

The commit adds a better handling for flash messages in `flashMessage.php` by implementing checks for undefined types, colors, and messages, and creating a script for the dynamic display of messages. It also refactors the login page in `login.php`- it improves styling and reorders elements for a cleaner layout, resulting in a more straightforward and user-friendly interface. The flash message component is now included at the start of the login page for better visibility.

* Update sidebar height style

Changed the height styling of the sidebar from fixed (`h-screen`) to minimum height (`min-h-screen`). This change will assure that the sidebar stretches depending on the content, instead of having a fixed height.

* Remove debug log and enhance modals

Debug log was removed from the 'pinjam' view. Detailed event modal now smoothly appears and disappears with the use of transform and opacity CSS properties. This change should improve the user experience when interacting with event details.

* Update user profile and password change layout

Incorporated the shared flash message into the user profile view. Adjusted design and dimensions of the profile and password change components for a more user-friendly layout. In the updated design, labels and input fields are more streamlined to ensure better alignment and consistency across the application.

* Update form UI and upload functionality

Overhauled the UI for the user forms, adding more interactive and flexible components, such as enhanced upload fields with detailed instructions and restrictions. Modified flash messages to enhance their visibility and nuanced them for a more localized experience. Also restructured upload functionality in the back-end to categorize allowed file types for document and image uploads separately.

* Refactor inbox view, centralize no-notification message

Applying code formatting and structure improvements to both user and admin inbox views. Also, updated the "no notification" message to be centrally aligned for better visual appeal. The changes enhance code readability and provide clearer user feedback when there are no incoming notifications.

* Refactor detailPeminjamanAcara modal view

The detailPeminjamanAcara modal view has been refactored and encapsulated within a function for cleaner code structure. It now features transition effects and improved visual hierarchy within its elements. The refactoring enhances the maintainability of the code and improves user interaction with the modal.

* Update user dashboard and detail room views

Added a session functionality for handling floor selection in the user dashboard view to improve data persistence. Additionally, refactored and enhanced the responsiveness of the detail room view. The changes improve UI consistency and enhance the usability of the application on varying device screens.

* Update CSS for responsiveness and UI consistency

Added and updated CSS classes to improve styling, layout, and responsiveness across different screen sizes. The updates also improve user interface consistency and usability across various device screens. Additionally, various non-relevant or unused CSS classes were removed to reduce clutter and streamline the stylesheets.

* refactor

* History (#55) (#56)

* Refactor login check into its own method

Moved the function for checking if a user is logged in into its own method. This is intended to enhance code readability and maintainability. It eliminates duplicated code blocks across multiple files, providing a single point of control for logic adjustments.

* Refactor history browsing function and template parameters

Separated display and data fetching concerns for user history browsing functionality. This heightens code readability and better follows software design principles. Also, optional chaining is added in item templates to ensure the code won't break due to missing or null data.

* Implement separate HistoryModel for user history browsing

The previous implementation had the history functionality within the PeminjamanModel. This commit extracts the history() function and places it into a new HistoryModel, thereby achieving a better separation of concerns. This change enhances the modular approach, improving the code's readability and maintainability.

* Refactor user history feature into separate HistoryModel

This commit moves the user history browsing feature from the PeminjamanModel to a newly created HistoryModel. This refactoring process improves the separation of concerns and promotes modularity, enhancing the code's maintainability and readability.

* Update CSS and enhance user history modal functionality

This commit includes various changes to the CSS file, notably removing redundant classes and adding new ones. It also improves the functioning of the user history modal in the user history PHP view. The modal now smoothly appears and disappears with a transition effect, improving the user experience.