Refactor modals to use functions and improve animation
Introduced functions to render modals for upload of Surat, detail of Mata Kuliah, and detail of Peminjaman Acara. The structure and styling of these modals were also updated to improve animation during the showing and hiding process. These changes enhance code reusability and provide smoother user interaction with the modals.