Implement detail course view with dynamic data
The changes primarily make room booking details dynamic, fetching data for the specific booked room with AJAX. This is achieved by adding a detail course method to the Peminjaman model, a handler method in the Request Controller and connecting them in the frontend with a click event listener. Relevant data is displayed in a modal box upon a button click providing an enhanced user experience.